search input: Move query invalidation to reusable module
This moves the query validation extension into its own module and adds
it to the new query input.
