+++ client/branded/src/search-ui/input/CodeMirrorQueryInput.tsx
@@ -2,13 +2,11 @@ import React, { RefObject, useCallback, useEffect, useMemo, useRef, useState } f
 
 import { closeCompletion, startCompletion } from '@codemirror/autocomplete'
 import { defaultKeymap, history, historyKeymap } from '@codemirror/commands'
-import { Diagnostic as CMDiagnostic, linter, LintSource } from '@codemirror/lint'
 import { EditorSelection, Extension, Prec, Compartment, EditorState } from '@codemirror/state'
 import { EditorView, ViewUpdate, keymap, placeholder as placeholderExtension } from '@codemirror/view'
 import classNames from 'classnames'
 import { useNavigate } from 'react-router-dom'
 
-import { renderMarkdown } from '@sourcegraph/common'
 import { TraceSpanProvider } from '@sourcegraph/observability-client'
 import { useCodeMirror, createUpdateableField } from '@sourcegraph/shared/src/components/CodeMirrorEditor'
 import { useKeyboardShortcut } from '@sourcegraph/shared/src/keyboardShortcuts/useKeyboardShortcut'
@@ -19,7 +17,6 @@ import {
     type QueryState,
     type SearchPatternTypeProps,
 } from '@sourcegraph/shared/src/search'
-import { Diagnostic, getDiagnostics } from '@sourcegraph/shared/src/search/query/diagnostics'
 import { appendContextFilter } from '@sourcegraph/shared/src/search/query/transformer'
 import { fetchStreamSuggestions as defaultFetchStreamSuggestions } from '@sourcegraph/shared/src/search/suggestions'
 import { RecentSearch } from '@sourcegraph/shared/src/settings/temporary/recentSearches'
@@ -28,8 +25,9 @@ import { isInputElement } from '@sourcegraph/shared/src/util/dom'
 
 import { createDefaultSuggestions, singleLine } from './codemirror'
 import { decorateActiveFilter, filterPlaceholder } from './codemirror/active-filter'
+import { queryDiagnostic } from './codemirror/diagnostics'
 import { HISTORY_USER_EVENT, searchHistory as searchHistoryFacet } from './codemirror/history'
-import { queryTokens, parseInputAsQuery, setQueryParseOptions } from './codemirror/parsedQuery'
+import { parseInputAsQuery, setQueryParseOptions } from './codemirror/parsedQuery'
 import { querySyntaxHighlighting } from './codemirror/syntax-highlighting'
 import { tokenInfo } from './codemirror/token-info'
 import { QueryInputProps } from './QueryInput'
@@ -535,83 +533,3 @@ const [callbacksField, setCallbacks] = createUpdateableField<
         }
     }),
 ])
-
-/**
- * Sets up client side query validation.
- */
-function queryDiagnostic(): Extension {
-    // The setup is a bit "strange" because @codemirror/lint only triggers
-    // linting when the document changes. But in our case the linting rules
-    // change depending on the query "type" (regexp, structural, ...). Changing
-    // the query type does not involve changing the document and to linting
-    // wouldn't be triggered. To work around this we explictly reconfigure the
-    // linter via a compartment when the parsed query changes but the document
-    // hadsn't change. This queues a new linting pass.
-    // See
-    // - https://discuss.codemirror.net/t/can-we-manually-force-linting-even-if-the-document-hasnt-changed/3570/2
-    // - https://github.com/sourcegraph/sourcegraph/issues/43836
-    //
-    const source: LintSource = view => {
-        const query = view.state.facet(queryTokens)
-        return query.tokens.length > 0 ? getDiagnostics(query.tokens, query.patternType).map(toCMDiagnostic) : []
-    }
-    const config = {
-        delay: 200,
-    }
-
-    const linterCompartment = new Compartment()
-
-    return [
-        linterCompartment.of(linter(source, config)),
-        EditorView.updateListener.of(update => {
-            if (update.state.facet(queryTokens) !== update.startState.facet(queryTokens) && !update.docChanged) {
-                update.view.dispatch({ effects: linterCompartment.reconfigure(linter(source, config)) })
-            }
-        }),
-        EditorView.theme({
-            '.cm-diagnosticText': {
-                display: 'block',
-            },
-            '.cm-diagnosticAction': {
-                color: 'var(--body-color)',
-                borderColor: 'var(--secondary)',
-                backgroundColor: 'var(--secondary)',
-                borderRadius: 'var(--border-radius)',
-                padding: 'var(--btn-padding-y-sm) .5rem',
-                fontSize: 'calc(min(0.75rem, 0.9166666667em))',
-                lineHeight: '1rem',
-                margin: '0.5rem 0 0 0',
-            },
-            '.cm-diagnosticAction + .cm-diagnosticAction': {
-                marginLeft: '1rem',
-            },
-        }),
-    ]
-}
-
-function renderMarkdownNode(message: string): Element {
-    const div = document.createElement('div')
-    div.innerHTML = renderMarkdown(message)
-    return div.firstElementChild || div
-}
-
-function toCMDiagnostic(diagnostic: Diagnostic): CMDiagnostic {
-    return {
-        from: diagnostic.range.start,
-        to: diagnostic.range.end,
-        message: diagnostic.message,
-        renderMessage() {
-            return renderMarkdownNode(diagnostic.message)
-        },
-        severity: diagnostic.severity,
-        actions: diagnostic.actions?.map(action => ({
-            name: action.label,
-            apply(view) {
-                view.dispatch({ changes: action.change, selection: action.selection })
-                if (action.selection && !view.hasFocus) {
-                    view.focus()
-                }
-            },
-        })),
-    }
-}
